[revert] "[dev] [JSON Schema] [experimental] replace custom parsers with JSON Schema and Ajv"
This reverts commit e7a833a39dff71c07176ee66891640eb8d5a1f1e.

This is kind of over-engineering most probably. The schema checking is
valuable and it was nice to drop some of the custom (albeit simple)
parsers for something extensible and closer to TypeScript's virtues. I
am sad there's not better support for deep merging, patching, and
defaults, more clarity with distinction from JSON-LD and
related technologies, and more maturity. The TypeScript back and forth
interface generation looks promising. For more sophisticated parsing
behavior like setText() I just call that in a layout pass. I still have
to deep mix the defaults which is a shame.

https://json-schema.org/
https://ajv.js.org/keywords.html
https://quicktype.io/typescript/
https://spin.atomicobject.com/2018/03/26/typescript-data-validation/
https://github.com/YousefED/typescript-json-schema
https://github.com/bcherny/json-schema-to-typescript
https://github.com/epoberezkin/ajv/blob/master/CUSTOM.md
https://github.com/epoberezkin/ajv-keywords/blob/master/keywords/prohibited.js
https://github.com/epoberezkin/ajv
https://github.com/epoberezkin/ajv/blob/master/KEYWORDS.md